+++ acvm-repo/acir/src/native_types/mod.rs
@@ -1,0 +1,8 @@
+mod expression;
+mod witness;
+mod witness_map;
+
+pub use expression::Expression;
+pub use witness::Witness;
+pub use witness_map::WitnessMap;
+pub use witness_map::WitnessMapError;