+++ compiler/integration-tests/test/integration/browser/compile_prove_verify.test.ts
@@ -1,4 +1,6 @@
 import { expect } from "@esm-bundle/chai";
+import { TEST_LOG_LEVEL } from "../../environment.js";
+import { Logger } from "tslog";
 import { initializeResolver } from "@noir-lang/source-resolver";
 import newCompiler, {
   compile,
@@ -14,6 +16,7 @@ const mnemonic = "test test test test test test test test test test test junk";
 const provider = new ethers.JsonRpcProvider("http://localhost:8545");
 const walletMnemonic = ethers.Wallet.fromPhrase(mnemonic);
 const wallet = walletMnemonic.connect(provider);
+const logger = new Logger({ name: "test", minLevel: TEST_LOG_LEVEL });
 
 const { default: initACVM, executeCircuit, compressWitness } = acvm;
 const { default: newABICoder, abiEncode } = abi;
@@ -24,7 +27,7 @@ await newCompiler();
 await newABICoder();
 await initACVM();
 
-compilerLogLevel("DEBUG");
+compilerLogLevel("INFO");
 
 async function getFile(url: URL): Promise<string> {
   const response = await fetch(url);
@@ -59,6 +62,9 @@ suite.timeout(60 * 20e3); //20mins
 
 test_cases.forEach((testInfo) => {
   const test_name = testInfo.case.split("/").pop();
+  const caseLogger = logger.getSubLogger({
+    prefix: [test_name],
+  });
   const mochaTest = new Mocha.Test(
     `${test_name} (Compile, Execute, Prove, Verify)`,
     async () => {
@@ -89,7 +95,7 @@ test_cases.forEach((testInfo) => {
       expect(noir_source).to.be.a.string;
 
       initializeResolver((id: string) => {
-        console.log("Resolving:", id);
+        caseLogger.debug("source-resolver: resolving:", id);
         return noir_source;
       });