+++ compiler/noirc_frontend/src/parser/parser.rs
@@ -184,6 +184,7 @@ fn function_definition(allow_self: bool) -> impl NoirParser<NoirFunction> {
                 is_unconstrained: modifiers.0,
                 is_open: modifiers.1,
                 is_internal: modifiers.2,
+                is_public: modifiers.3,
                 generics,
                 parameters,
                 body,
@@ -199,13 +200,14 @@ fn function_definition(allow_self: bool) -> impl NoirParser<NoirFunction> {
 /// function_modifiers: 'unconstrained'? 'open'? 'internal'?
 ///
 /// returns (is_unconstrained, is_open, is_internal) for whether each keyword was present
-fn function_modifiers() -> impl NoirParser<(bool, bool, bool)> {
+fn function_modifiers() -> impl NoirParser<(bool, bool, bool, bool)> {
     keyword(Keyword::Unconstrained)
         .or_not()
+        .then(keyword(Keyword::Pub).or_not())
         .then(keyword(Keyword::Open).or_not())
         .then(keyword(Keyword::Internal).or_not())
-        .map(|((unconstrained, open), internal)| {
-            (unconstrained.is_some(), open.is_some(), internal.is_some())
+        .map(|(((unconstrained, public), open), internal)| {
+            (unconstrained.is_some(), open.is_some(), internal.is_some(), public.is_some())
         })
 }
 
@@ -1416,10 +1418,6 @@ where
         .then(expr_no_constructors.clone())
         .map(|(start, end)| ForRange::Range(start, end))
         .or(expr_no_constructors.map(ForRange::Array))
-        .validate(|expr, span, emit| {
-            emit(ParserError::with_reason(ParserErrorReason::ForLoopDefaultTypeChanging, span));
-            expr
-        })
 }
 
 fn array_expr<P>(expr_parser: P) -> impl NoirParser<ExpressionKind>